Draw tildes and a welcome message
There is a new function got introduced here in this commit. And that is
v_rfsh_scr(). This function will act as a main way of refreshing the
entire editor screen. Every single terminal line drawing will be done
here, and yes, it heavily relies on NCURSES (or curses if you're a
dinosaur) functions to do the work like any other part of the editor.

Signed-off-by: Luth <71706257+mkluth@users.noreply.github.com>