File I/O support (reading operation)
Like any other text editors out there, the ability to open, read and
write to a file is a necessary thing to have. Or else, we can't even
call it a text editor in the first place if it cannot even do things
like basic file editing. So in order to achieve this, lets implement
the ability to open and read to a file first. Rather than reading the
entire file content and display them, lets starts doing things slowly.
For now, lets read a single line from a file and display its content.

Signed-off-by: Luth <71706257+mkluth@users.noreply.github.com>