Support horizontal scrolling
Its not like the editor does not supports the horizontal cursor movement
thing before. It does. However, we can't move the cursor to the right
edge of the screen to view the existing text that cannot be displayed
within the terminal window (due to its lengthy length). This commit
shall fix this problem for us and forever.

Signed-off-by: Luth <71706257+mkluth@users.noreply.github.com>