src/input.c: Limit cursor scrolling to the right and snap it to EOL
We don't want the user from moving the cursor way off to the right of a
line and starts inserting text there. It doesn't make sense. However,
there is an exception, where the cursor X position *can* point one line
past the EOL so that new lines can be easily added there. Ah also, there
is a need for us to force snap the cursor to EOL. The reason behind, is
due to the fact that the user can still move past the EOL, even though
we already implement the scrolling limitation stated above. This can be
done by moving the cursor to the end of a long line inside the text file
before moving the cursor down to the next line (only works if the next
line is shorter than the previous one). But still, it doesn't make sense
either way. So, lets patch that up.

Signed-off-by: Luth <71706257+mkluth@users.noreply.github.com>