Add tabs rendering
Don't get me wrong. Tabs are rendered pretty nicely in the curses mode
currently. There's no characters displayed within a tab; despite the
fact that a tab key only pushes the cursor to the next tab stop, leaving
the characters within it untouched. However, the same does not applied
to the cursor. The way the cursor interacts with a tab character is
pretty much *ugly*. Like, it doesn't even jump to the next tab stop like
how it should be like any text editors out there. The reason behind, is
due to the fact that we only manipulate the cursor position within
the screen rather than within the line itself (yep, surprise!). We need
to fix this somehow. But for now, lets render the tabs by ourselves
first instead of leaving it to the hand of NCURSES.

Ah also, in order to render a line, we need to have a new string that
contains a rendered version of the original text. We can use this string
later on in the future to display nonprintable characters such as the
Ctrl keys. But for now, lets just focus on the tab key.

Signed-off-by: Luth <71706257+mkluth@users.noreply.github.com>