Make the cursor interact nicely with the rendered tabs
Alright, we already done the line rendering part on the previous commit.
However, the cursor still doesn't seems to interact nicely with it.
Well, the reason behind, is due to the fact that we still assuming each
character takes up only one column while we're positioning the cursor.
To fix this, we need a new x-axis variable that tracks the cursor
position on the rendered line instead off on the actual original line
itself. We will use this new variable to manipulate the cursor behaviour
on that line a.k.a "rendering" it.

Signed-off-by: Luth <71706257+mkluth@users.noreply.github.com>