Display simple and short informations on the status bar
Yeah, the status bar now already have some informations displayed on
it. The informations displayed for now only includes the currently
opened filename, and the total number of lines available in the text
file. I also added a fallback handling if somehow the user terminal
does not supports colors manipulations. Not to mention, I also changed
the editor welcome message a little bit. Nothing much.

Also, I don't know why but the text displayed on the highlighted bar
doesn't seems to get displayed properly. It should be seen as a bold
black text. However, my terminal displays it as a bold grey text
instead. Weird. I'm not sure about your terminal though. But I think it
could be something to do with the editor codebase, or fear not, the
NCURSES library itself. I'm not too sure about this anyway.

Signed-off-by: Luth <71706257+mkluth@users.noreply.github.com>