Display more informative and useful details on the status bar
Rather than only displaying the filename and the total number of lines,
let's display something a lot more informative and useful. Yes! That
would be the current line number where the cursor is currently at, and
of course, which column in that line the cursor is currently located.
Also, let's display the current status of the file editing too. By what
I meant here is that, does we have any unsaved changes? If yes, then
let's tell the user that there are unsaved ones by showing a symbol on
the status bar itself. I also made some improvements on the
documentations in this commit (especially on the header file). That's
pretty much the sum of this.

Signed-off-by: Luth <71706257+mkluth@users.noreply.github.com>