Unbold the status bar text
I already decided that it is much better for me to leave the status bar
text goes unbold. The reason is that, I don't know why, it's just the
bold text doesn't go well with the status bar color pair (which is black
for the foreground, white for the background). However, I do keep the
bold status bar text for cases where the user's terminal doesn't support
curses colors attributes either way. I'm not sure if all terminals gonna
support the bold text, I might be on the wrong side here. But hey, let's
keep it. If it's not gonna work, well, we'll remove it later on.

Also, I decided to just use the default color macros defined by the
NCURSES library for the status bar color definitions (I used their
numeric values before) instead. A lot easier to read and understand.

Signed-off-by: Luth <71706257+mkluth@users.noreply.github.com>