Add a status message feature
A status message shall be displayed below the editor's status bar
starting from this commit. This will be used for displaying messages to
the user and prompting for input. This will be really useful once we
started implementing the editor's search feature later in the future.
However, do note that a status message will have a timeout depending on
the value of V_STATS_MSG_TIMEOUT macro. Once the timestamp of the
message exceeded the value of the timeout macro, the message shall be
erased. By default, the timeout is setted for 5 seconds. I might change
this in future, who knows? Unlike most strings in editor source code
which are commonly dynamics, the string used for the status message
isn't.

The string is static and the maximum value depended on the value of
V_STATS_MSG_BUF macro. Checkout the void.h header for more details.
Same thing with the timeout feature, I might change this part out later
in the future. Just wait and see. I might make the string to be dynamic
next time.

Signed-off-by: Luth <71706257+mkluth@users.noreply.github.com>