Allows changes made to be saved to disk
Finally we can now save the changes made by the user directly to disk!
So here's the analogy behind. Firstly, the editor would read the entire
v->rows array before converts them into one long ass string. Once the
conversion is done, we will then open the file we wanted to save before
we dump out the long ass string inside it and close it. Do note though,
this editor is only targeted on Linux and macOS because I used POSIX
functions for the saving part. Unfortunately, this saving feature only
works if the file already existed. Wait for the next upcoming commit.

Signed-off-by: Luth <71706257+mkluth@users.noreply.github.com>