Add dirty flag to warn the user
Let's keep track of whether the text loaded in the editor differs from
what's in the file. From here, we can warn the user that they might lose
changes they made if they try to quit without saving. If there are some
unsaved changes available, the symbol '[+]' will appear on the status
bar. The symbol will be gone once the user pressed Ctrl-S, which meant
they saved the file.

Signed-off-by: Luth <71706257+mkluth@users.noreply.github.com>