Add quit confirmation
Before we exit the editor, let's check if there are any unsaved changes
that were made by the user. If we did found it, display a warning
message as a status message and ask for the user confirmation. In order
to confirm, the user must hit Ctrl-Q one more time and the editor shall
exits without saving any changes on the file. That should be convenient
enough.

Signed-off-by: Luth <71706257+mkluth@users.noreply.github.com>