Add a simple backspacing
Okay, now we can actually backspacing any mistakes that we've done.
However, do note that this backspacing feature for now only works for
backspacing the character on the left. Backspacing at the start of a
line isn't a thing yet, the same thing applied to backspacing on the
right side and backspacing the character located underneath the
cursor. I shall release a commit that will handle all of these in the
future, so no worries! At least, we're now a lot friendlier for the
user than ever before.

This commit also fixed several docstrings documentation mistakes in
src/editor.c and also the bug where there will be a short delay after
the user pressed the ESC key when switching back to the editor
Command Mode. I basically just set the ESC key delay to 0. It's not
really a bug but rather the way ncurses originally handles keypresses
from getch(). So yeah, fuck it. Let's handle it our own way instead.

Signed-off-by: Luth <71706257+mkluth@users.noreply.github.com>