Makefile: Use the default C compiler according to $(CC)
Rather than hardcoded which compiler to use, let's just leave it to the
hand of the system's symbolic link and just compile with whatever the
default C compiler is. I also changed the position of -lncurses
argument to come after the object files when linking, allowing the
linking command to be compatible with both GCC and Clang.

Signed-off-by: Luth <71706257+mkluth@users.noreply.github.com>