Discontinue the support for automatic status message erasing feature
I don't think there's a need for us to support the status message
timestamp for the time being. Plus, we can't really erase off the
status message once it reached its maximum timeout automatically if the
way we get the user's input itself is through a blocking function call
(the major issue here is that we used the plain getch() function from
the NCURSES library to get the user's input). I will try to find the
alternatives and workarounds for this blocking function problem in
the future, but for now, let's forget this feature for a while.

For the time being, we have to *manually* erase off the status message.
Sorry.

Signed-off-by: Luth <71706257+mkluth@users.noreply.github.com>