Allow cursor movement at the BOL and at the EOL
Let's make the cursor movement a lot freely. Starting from this commit,
pressing left at the BOL (Beginning of Line) would pushes the cursor
backwards to the EOL (End of Line) of the previous line. Meanwhile,
pressing right at the EOL would pushes the cursor forward to the BOL
of the next line.

Much better yeah?

Signed-off-by: Luth <71706257+mkluth@users.noreply.github.com>