Add Enter key support (yes, finally)
Yes, you hear it right. Now the editor can be fully use as a basic text
editor by this commit. But yeah, in the end it's a plain piece of editor
and it's quite unstable to be honest, as there are too many unknown bugs
that I, myself not even encountered with before (but sooner or later they
shall appear too either way). At least for now, Void can be used for some
simple and quick text editing purposes. I won't say it's a great tool
because it just too darn basic and unstable. So, use it at your own risk
(nobody gonna use this anyway and even I myself will not use it as a daily
text editor for my work).

This commit also includes some docstrings updates for include/void.h and
that's it. :)

Signed-off-by: Luth <71706257+mkluth@users.noreply.github.com>