Handle Vim-based keys (0 and $) and GNU Nano Page Up/Down shortcut keys
Pressing 0 in Command Mode makes the cursor behaves the same way as
pressing the Home key. Meanwhile, pressing $ makes the cursor to
behaves the same as by pressing the End key. Pressing Ctrl-V on the
other hand is the same thing as pressing the Page Down key. Pressing
Ctrl-Y is the same thing as pressing down the Page Up key. That's pretty
much it for this commit. I just wanted to make the editor a lot easier
for me to navigate with (thanks to my experiences of using Vim for such
a long time). That's why I decided to roll out this commit.

Signed-off-by: Luth <71706257+mkluth@users.noreply.github.com>