Rewrite the input handling + just use the stdscr
I decided to change the way we handle input keys in this commit. As a
result, this commit forces the editor to only supports one input key for
now and that is the Ctrl-Q (for the sake for us to quit from the editor).
I also realized that there's no need for us to have v_win in the editor
v_state, its useless. Its way better for us to just use and rely on the
stdscr provided by NCURSES library instead, makes the code much simple
and clearer.

I did try to run this program binary inside Valgrind and it seems like
there are leaks happening behind (thx NCURSES) but the great news is
that there are no leaks happens at our side as the developers and that's
a great news! Unfortunately, nothing we could do about the leaks coming
from NCURSES anyway, so gotta let it slips :/

Signed-off-by: Luth <71706257+mkluth@users.noreply.github.com>