Make the status bar text displays on the right side too
I personally think that having everything displayed on the left is just
too uncomfortable for me (especially when editing a file for a long
time). So, I decided to make the cursor x-position and y-position to be
printed on the right side of the screen instead. Its a lot easier on the
eyes too, don't ya think?

Signed-off-by: Luth <71706257+mkluth@users.noreply.github.com>