Add LICENSE for the repository
Let's go with MIT one, okay? For those who wanted to use this
editor, please keep this in mind. I will not responsible for
any losses you faced while using this editor because of bugs or
whatever the problem is. This project is just a toy project I
started years ago and I never once imagine that this stupid
editor I created will be use widely in the future by someone or
whatsoever. So yeah, if you still insisted of using void, you
got balls and I respect your decision a lot!

P/S: Feel free to contribute to this project. I'll look forward
for those of you who are interested :)

Signed-off-by: Luth <71706257+mkluth@users.noreply.github.com>