Add SIGWINCH handling and filter out user input before v_insert()
Yes, the editor now can easily handles SIGWINCH (Signal Windows Change)
without any problem. Users can easily resize the editor window without
worrying about breaking the editor graphical layout. But, there's a catch
since I'm using a simple blocking getch() for this editor. The editor
screen refreshing process won't happen immediately, and it is necessary
for the user to actually type out any random key to actually force the
editor to refresh itself (if the key is printable then it shall be printed
onto the screen AND into the currently opened buffer -- except if you're
in Command Mode and the key itself is not mapped to any action at all).

I was told by ChatGPT that this behavior is actually pretty classic for
any terminal programs that uses curses mode with a blocking input
function. So, I guess I'll take that. It's not like I never tried to
implement an unblocking getch() before. I did. But, the CPU usage spiked
up and I hate to see it and acknowledge that, so I keep the blocking one
instead even until now (I CAN tune it down so that it uses less CPU, but
nevermind, maybe you can tune it for me instead, how about that? Hehehe).

I also filter out user input before v_insert() is called because
whenever you're resizing in Insert Mode, this stupid ~Z will be printed.
So, I patched that up. Let's hope this one works on Termux too and also
on your machine. Adios.

Signed-off-by: Luth <71706257+mkluth@users.noreply.github.com>